Organizes Acts in /Acts
Organizes the Act Components into /Acts
